refactor(phase6): unify color map, move value constants from types to lib/constants
ST6-1:
- Delete getColorConfig from lib/utils.ts (dead code — zero callers)
- PRIORITY_CONFIG and IMPORTANCE_CONFIG share single LEVEL_CONFIG source (values were identical)

ST6-2:
- Move DEFAULT_KANBAN_COLUMNS, KANBAN_COLUMNS, PRIORITY_CONFIG, IMPORTANCE_CONFIG,
  KANBAN_STATUS_COLORS, TAG_COLORS, initialFilterState → lib/constants.ts
- Remove all export const from types/index.ts (types-only file after this change)
- Update 10 importer files to import values from @/lib/constants

All 442 tests pass, TypeScript clean.
